change shortcode logic to take generate from a counter

+++ src/main/java/com/mohit/urlshortener/util/Base62.java
@@ -4,10 +4,10 @@ public class Base62 {
 
     private static final String BASE62 = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
 
-    public static String encode(int num) {
+    public static String encode(long num) {
         StringBuilder sb = new StringBuilder();
         while (num > 0) {
-            sb.append(BASE62.charAt(num % 62));
+            sb.append(BASE62.charAt((int) num % 62));
             num /= 62;
         }
         return sb.reverse().toString();
